refactor(sensor_board): :recycle: renamed logger.c/h -> data.c/h
